Created new test case for configuration end point. Did not cover UserAPI endpoint due to AMBR-175. Renamed files for simplicity.
